feat(server): resend confirmation email when trying to create account with existing email
But don't resend email if user has already confirmed their account!
